requirements.txt
pydantic==1.8.2
requests==2.26.0
pandas==1.3.0
pyyaml==5.4.1
git+https://github.com/tweepy/tweepy.git@145d7cfb902ed7bd10143e201ea89af41ba7d628
git+https://github.com/dhudsmith/twitter-alchemy.git@739fa4423534276f4208256c946e57fc238e0fbf